+++ demo/gulp/styles.js
@@ -13,7 +13,7 @@ gulp.task('styles', function () {
   };
 
   var injectFiles = gulp.src([
-    paths.src + '/{app,components}/**/*.scss',
+    paths.src + '/app/**/*.scss',
     '!' + paths.src + '/app/index.scss',
     '!' + paths.src + '/app/vendor.scss'
   ], { read: false });
@@ -21,7 +21,6 @@ gulp.task('styles', function () {
   var injectOptions = {
     transform: function(filePath) {
       filePath = filePath.replace(paths.src + '/app/', '');
-      filePath = filePath.replace(paths.src + '/components/', '../components/');
       return '@import \'' + filePath + '\';';
     },
     starttag: '// injector',