+++ demo/gulp/watch.js
@@ -7,8 +7,8 @@ var paths = gulp.paths;
 gulp.task('watch', ['inject'], function () {
   gulp.watch([
     paths.src + '/*.html',
-    paths.src + '/{app,components}/**/*.scss',
-    paths.src + '/{app,components}/**/*.js',
+    paths.src + '/app/**/*.scss',
+    paths.src + '/app/**/*.js',
     'bower.json'
   ], ['inject']);
 });